+++ src/api/preferences/index.ts
@@ -1,0 +1,3 @@
+export * from "./deletePreferences";
+export * from "./getPreferences";
+export * from "./updatePreferences";